+++ undersampling/opf_us2.py
@@ -12,7 +12,7 @@ class OpfUS2(US):
 	def variant(self, output, X, Y, majority_class, minority_class):
 		#2st case: remove samples from majoritary class with negative or zero scores
 		output_majority = output[output[:,1]==majority_class]
-		output_majority_neutal = output_majority[output_majority[:,2]<=0]
+		output_majority_neutal = output_majority[output_majority[:,2]<=0].astype(int)
 
 		X_train = np.delete(X, output_majority_neutal[:,0],0)
 		y_train = np.delete(Y, output_majority_neutal[:,0])
