+++ undersampling/opf_us3.py
@@ -18,7 +18,7 @@ class OpfUS3(US):
     
 	def variant(self, output, X, Y, majority_class, minority_class):
 		#3st case: remove all samples with negative
-		output_negatives = output[output[:,2]<0]
+		output_negatives = output[output[:,2]<0].astype(int)
 
 		X_train = np.delete(X, output_negatives[:,0],0)
 		y_train = np.delete(Y, output_negatives[:,0])
